Use the l2chroot var to avoid relying on PATH
This allows l2chroot to be installed somewhere that is not on the PATH
(as the var has an absolute path). It also allows for renaming the
l2chroot script to allow multiple chroots on a system without worrying
about conflicting changes to the same script.

The same change was also implemented here:
https://github.com/gauch/ansible-role-ssh-chroot-jail/commit/4a9a34288af0f739adad5dec22c19e513ff4e10e
https://github.com/paradigmadigital/ansible-role-ssh-chroot-jail/commit/4238e4f450f6f42b823def38cefbd9eaeb19a2e5